Fix client-side sorting after adding a column
With arrival of NetHack 3.6.0 we added a column
into tables in the 'nh' variant. This broke the
client-side sorting. This commit makes the
sorting plug-in configured from the custom
attributes in TH table elements. The .js file
was switched to Unix line-endings in the process,
which is why the whole file is a change :-(