Change behavior of 'static' flag with logfiles table entry
Logfiles table defines the sources of data for NHS; if a row in the table
has 'static' flag, the successfull processing will be followed by
flipping the 'oper' flag to false, which will cause all next runs of the
feeder to skip that entry.

Therefore, marking a source as static (with the --static option) will
safely a) process outstanding data, b) mark the row to never be processed
again.