Regularize logfiles.dumpurl semantics
logfiles.dumpurl contains the template that is used to construct game
dump links.  For NetHack variants based on NH4 this field was handled in
different, inconsistent way: It only contained the path to which the
actual dump filename was appended.  This commit changes this semantic by
adding %d and %D template tokens that resolve into the game dump filename
(the %D form additionally replaces _ with :), making the usage consistent
for all variants.

So, following dumpurl specification for NetHack4:

    https://ascension.run/userdata/%u/nethack4/dumplog/

Will become:

    https://ascension.run/userdata/%u/nethack4/dumplog/%D