Change endtime field in the database views
The 'endtime' field is changed to contain the native database
representation of the end time (TIMESTAMP WITH TIME ZONE), the formated
version is in 'endtime_fmt' (UNIX epoch format remains unchanged in
'endtime_raw').

This is done to have access to the native field for date/time operations
without unnecessary type casts.