Implement a way of signalling source options
Sources defined in "logfiles" table now have new field "options" that is
an array of textual tags that signal something special should be done
about the source (for example, trigger a workaround for a bug).

Currently the only tag recognized is "bug360duration" that should be used
with unpatched NetHack 3.6.0 which has a bug where bogus 'realtime' field
is being written to xlogfile.  This was formerly handled by
special-casing in the feeder and had the disadvantage of not allowing
accommodation of patched 3.6.0 sources.