Remove all /dev/null/nethack code
The last /dev/null/nethack Tournament took place in November 2016 and the
official retirement was announced in September 2017.  This is significant
event for the NetHack Scoreboard and me personally.

The NetHack Scoreboard grew from the alternate scoreboard page that I
hurriedly cobbled together in November 2013.  It was a simple supplement
to the original very slow and very crummy scoreboard that allowed viewing
the basic information like all ascensions, basic player information etc.
It was simply a stats script that read the devnull's xlogfile and spit1
out html pages.  It's worth mentioning that the web design, "borrowed"
from DCSS's [akrasiac
scoreboard](http://crawl.akrasiac.org/scoring/overview.html) is still
basically unchanged from that time.

Soon after /dev/null/nethack 2013 ended I started to work on universal
NetHack scoreboard that would aggregate multiple public sources in one
place.  To continue to suport /dev/null/nethack tournament the new
scoreboard included code specific to it.  In fact quite a lot of code.
That's the code that is being removed with this commit.  That's the code
whose result is still possible to see
[here](https://scoreboard.xd.cm/devnull/).  Regenerating these pages will
from now only possible by rewinding the NetHack scoreboard to time before
this commit.

So, this is it. Goodbye, /dev/null/nethack. It was an unforgettable ride.