Implement some more flexibility to accomodate SLASH'EM Extended
In SLASH'EM Extended we don't know what roles/races exist, so we allow
everything and do not do any aggregation and tabulation based on these.
This affects also z-scores, that need some special-casing for SLEX, which
only list roles that have ascensions.  This might still eventually lead
to too wide tables, but we do not expect to see too many SLEX ascensions.