Fix a vulnerability where user can fake their ascension
Users are unfortunately able to put xlogfile field separator in monster
names and item names, which in turn leads to exploits where putting in
death=ascended is parsed as ascension by NHS.

This fix is really more of a workaround to prevent abuse. The proper fix
would be to not allow users to put field separator character into
anything that ends up in xlogfile.