Fix broken Conducts-all page
https://nethackscoreboard.org/conducts.all.html was showing 100
ascensions with all conducts broken, rather than a descending list
of high-conduct ascensions. Fix this by including assertions to
SQL query (present for variant-specific lookups) to the logic for
'all'.

When fixing this it was also noticed that for all or nh, some
ascensions are displayed in the incorrect order (by nconduct).
This seems to arise because the exact value of ncond is modified
by row_fix(), otherwise the order of results given by the SQL
query is correct. Will add logic to re-order after row_fix()
processing. That worked.

Fixup: original fix broke the rank order. Add code to also update
$row->{n} for each row after re-ordering. All good!

Reported-by: spicycat/#nhs-dev/Freenode on 30.07.2020
Signed-off-by: Joanna/aoei