Create init.sh script to automate docker build and run process
init.sh prepares configuration files that are hidden from the git tree
and runs the docker build process with docker-compose to generate images
and containers necessary to run the containerised dev version of the site.
It assumes a clean git tree and no pre-existing nhs docker containers.

Add an additional script, reloader.sh which can rebuild and reload various
components of an already built or partially built set of docker images
and containers for running NHS. This should rarely if ever be necessary,
in most cases restarting the nhs-web-m container should be sufficient.
It should also be possible to test changes to the front-end code without
having to reload the nhs-web-m container (running morbo) at all.

Signed-off-by: Joanna/aoei

Make more comprehensive init script for building and running containers

Added a lot of bells and whistles to the init script.