Allow registering AceHack games where mode='solo'
nhdb-feeder.pl was filtereing games for any variant with a 'mode' field in the xlog,
unless mode = 'normal' or 'challenge'. This unreasonably filters AceHack games with
mode = 'solo', which only disables bones and prevents the player from rerolling their
character as much as they like. nht had a 5k junethack AceHack game that was filtered.
This turned out to be an easy fix, just add a condition for mode = 'solo' to
nhdb-feeder.pl:516. For the live site, it will be necessary to re-import the AceHack
xlogfile data.
Reported by and signed-off-by: aoei/Joanna

fixup/readability: change $l to $xlog_line in nhdb-feeder.pl in main file read loop,
or to $xlog_data in sub sql_insert_games, also in the main file reading loop change
$pl to $parsed_line