Copy updated nethack_def.json from legacy/feeder to mojo
This is the downside of having the same file in two different places.
hard or symbolic links might be possible instead, but I was worried
about how that'd work out with container bind mounts. Actually,
pretty sure I tried symlinks at some stage and it was, unsurprisngly,
a no-go.

Want to think a bit more about desgin but ideally there should be
something like a simple Dockerfile for each container, to build the
image (might not even be necessary with the readily available postgres
container), and a single build script/config file combination that
will largely be maintained in the repository, with the exception of
database passwords, or other tokens that can be generated at build
time that only the software running the containers need.

I think moving a lot of NH/variant game rules and definitions to
the database may help simplify things. That's where the shared state
between the web server and the backend updater/aggregator is anyway.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>