Modify the behaviour of the high-conducts page
Issue was that the database counts conducts differently than the
scoreboard (ultimately this should also be brought in line), so
some higher-conduct games were not reaching the top 100 as displayed
by the scoreboard, and prior to commit 66a7ec6 some were also
ordered incorrectly.

In the database achievements are also included by bitcount, and
the page generation code gets the correct number of conducts by
post-processing. This additional, somewhat crude fix requests a
larger number from the database (200), and then returns 150 in
the table - 150 not only 100 because, why not?

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>