Fix warning about splicing too-small arrays
Conducts page sometimes complained as the last commit fixing the
out-of-order conducts page always tried to splice 50 off the end of
an array of 200.
Add bounds checks.
Also improve container run-script a bit. Should perhaps write an
outline on general usage. This commit also enables the first-to-asc
page for SpliceHack! (now spl) I hope it's actually working properly,
might not be immediately obvious if there are mistakes with some
roles/races etc.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>