UNSTABLE: Refactor tree
VCS is a mess when the source repo is also a runtime
directory for the software (even if you do use .gitignores).

I also think it's nice to have things a bit more logically
separated, as well as a clear system for config, build, run.

New structure is the following:
bin/    --  executable programs
defs/   --  default configurations and JSON definitions
lib/    --  perl modules required by runtime scripts
run/    --  run scripts + runtime env
pods/   --  Dockerfiles and initial state for containers
www/    --  Static content for the website (css, images)

defs will be used to generate cfg, run/* will be copied
to e.g. ~/nhs-run/, cfg and lib will also be copied there.
www would be copied to a webserver root, or bind-mounted
to a container.

nhdb-feeder.pl and nhdb-stats.pl will be run from ~/nhs-run,
or other installed directory, and expect to find lib and cfg
in there. If containerised, this dir on the host system
would be bind-mounted to /run/nhs inside the container.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>