Fix missing subdomain for TNNT dumplog URLs
nhdb-feeder.pl began to fail processing of current, live TNNT 2020
xlogdata, complaining about a NULL version field and halting updates.
The version field was certainly present in the xlogfile entries.

However, it was also noticed that TNNT dumplog URLs were broken links.
As specified by the dumpurl field for the respective entries in the
logfiles table, the appropriate subdomain (www for hdf-us, eu for
hdf-eu, or au for hdf-au) was specified with the token %S.

Some provision was already made for this token in the subroutine
url_substitute (in lib/NHdb/Utils.pm). However, the relevant 'src' field
in the merged xlogfile was not processed by nhdb-feeder.pl, nor saved
anywhere in the database. As a result, the %S token was falling back on
the default of an empty string.

Thus, to fix this issue, add a check to the subroutine sql_insert_games
in nhdb-feeder.pl. The check explicitly adds the src field, if present.
At that stage it is entered as-is (either hdf-us, hdf-eu or hdf-au).
Also explicitly add this optional src field to the games table defined
in schema/games.sql, as well as its views.

[Just as a side note, a more general approach was suggested previously by
Mandevil. This would be to include a misc field containing JSON data for
any unexpected xlog fields. One benefit of this would be that the
tournament special tnntachieve fields would also be retained in the NHS
database. This may be a nice idea in future, but I digress.]

Also add some additional processing to url_substitute, to convert the
src string from its original format hdf-(us/eu/au) to the correct
subdomain (www/eu/au). In local testing the dumplog URLS now appear to
work. I did not reproduce the the NULL version string error, either.

Lastly, update TNNT logurls for TNNT 2018/2019 to their archived
addresses. Also correct a source entry that erroneously used the domain
name hfa.hardfought.org instead of au.hardfought.org (logfiles_id 60,
au.hardfought.org NetHack 4).

Hopefully everything runs smoothly on the live site.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>