Prototype REST API queries: ascended and realtime
Very rough initial mockup for actix-web REST API.

Duplicated code in each page needs to be refactored into a single method
that unpacks and re-serializes game records as JSON.
Nullable int types (such as for starttime_raw and endtime_raw) need to
be handled with Option<T>, to avoid panics. Calls to unwrap() also need
to be replaced with proper error handling for db connection and query.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>