Define postgres types, refactor into gamedata.rs
REST API prototype requires a change to the DB scheme, we want a field
conducts that is an array of varchar types, each one being a string
representation of a conduct. Handling of bitfields for older NetHack
records will be moved to being a job for the aggregator.

We also now expect dumplog to be stored in the DB for us, which
currently isn't the case, although the field already exists.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>