Update conduct field int -> bigint
aosdict's changes involving the upper bits of conduct broke stuff
because postgres integer is signed and he put stuff in the sign bit.

The live db was modified to acommodate this by changing conduct to
a bigint.

TODO: conductX handling
ALSO TODO: versioning of bitfields for pre-conductX games, as the
conducts are still reported incorrectly for various version/variant
combinations.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>