Add nh37-setseed-hdf to scoreboard
In order to include these games I added an additional boolean field to
the main games table, and decided to back up the db and recreate the
whoel table from scratch.

The feeder script does not exclude games where user_seed=1, but flags
them appropriately. Games with user_seed=0 or with no user_seed field
will not be flagged as seeded.

Games and ascensions that are flagged as seeded are excluded from most
views for now, with the exception of v_games_all, which includes both
scummed and user-set seeded games. IMO, seeded games where the game
chose a random seed at the beginning of the game are just as valid as
vanilla ascensions that aren't seeded at all.

TODO: add a page/pages for displaying ascensions with user chosen seed

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>